Refactoring Java tests: 1-17
- Refactor Java 6 tests: improve dynamic compilation and add assertions for class accessibility
- Adjust wait time in Java 5 concurrency test
- Correct file path in Java 17 deserialization test
- Remove redundant whitespace in Java 4 file read test
- Refactor functional interface declaration in Java 8 tests

+++ JavaReleases/src/test/java/pl/mperor/lab/java/Java6.java
@@ -2,6 +2,7 @@ package pl.mperor.lab.java;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import pl.mperor.lab.common.TestUtils;
 
 import javax.script.ScriptEngine;
 import javax.script.ScriptEngineManager;
@@ -11,6 +12,10 @@ import javax.tools.ToolProvider;
 import java.awt.*;
 import java.io.File;
 import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
 import java.nio.file.Files;
 
 /**
@@ -32,12 +37,12 @@ public class Java6 {
     }
 
     @Test
-    public void testJavaDynamicCompilation() throws IOException {
-        File sourceFile = new File("./HelloWorld.java");
+    public void testJavaDynamicCompilation() throws IOException, ClassNotFoundException, InvocationTargetException, NoSuchMethodException, InstantiationException, IllegalAccessException {
+        File sourceFile = new File("HelloWorld.java");
         String javaSourceCode = """
                 public class HelloWorld {
-                    public static void main(String[] args) {
-                        System.out.println("Hello World!");
+                    public void sayHello() {
+                        System.out.print("Hello World!");
                     }
                 }
                 """;
@@ -45,11 +50,23 @@ public class Java6 {
 
         JavaCompiler compiler = ToolProvider.getSystemJavaCompiler();
         Assertions.assertNotNull(compiler, "JavaCompiler should not be null");
-
         int compilationResult = compiler.run(null, null, null, sourceFile.getAbsolutePath());
         Assertions.assertEquals(0, compilationResult, "Compilation should succeed with result 0!");
+        assertHelloWorldClassAccessible();
+
         Assertions.assertTrue(sourceFile.delete(), "Source file should be deleted after compilation");
-        Assertions.assertTrue(new File("./HelloWorld.class").delete(), "Now compilation result can be deleted!");
+        Assertions.assertTrue(new File("HelloWorld.class").delete(), "Now compilation result can be deleted!");
+    }
+
+    private void assertHelloWorldClassAccessible() throws MalformedURLException, ClassNotFoundException, NoSuchMethodException, InvocationTargetException, InstantiationException, IllegalAccessException {
+        URL[] urls = {new File("").toURI().toURL()};
+        URLClassLoader loader = new URLClassLoader(urls);
+        Class<?> clazz = loader.loadClass("HelloWorld");
+        Object instance = clazz.getDeclaredConstructor().newInstance();
+        var readableOut = TestUtils.setTempSystemOut();
+        clazz.getDeclaredMethod("sayHello").invoke(instance);
+        Assertions.assertEquals(readableOut.all(), "Hello World!");
+        TestUtils.resetSystemOut();
     }
 
     @Test
